Rename folders, add commission.module
