+++ src/app.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule, ConfigService } from 'nestjs-config';
 import * as path from 'path';
-import { CommissionController } from './controller/commission/commission.controller';
-import { CommissionService } from './service/commission/commission.service';
+import { CommissionController } from './controllers/commission/commission.controller';
+import { CommissionService } from './services/commission/commission.service';
 
 // ConfigService.rootPath = path.resolve(__dirname, '..');
