+++ src/config/db.ts
@@ -6,8 +6,8 @@ export default {
     port: process.env.DB_PORT,
     username: process.env.DB_USER,
     password: process.env.DB_PASSWORD,
-    entities: [path.resolve(__dirname, '../entity', '**/!(*.d).{ts,js}')],
-    migrations: [path.resolve(__dirname, '../migration', '**/!(*.d).{ts,js}')],
+    entities: [path.resolve(__dirname, '../entities', '**/!(*.d).{ts,js}')],
+    migrations: [path.resolve(__dirname, '../migrations', '**/!(*.d).{ts,js}')],
     database: process.env.DB_NAME,
     synchronize: false,
 };
