--- src/repository/TestRepository.ts
@@ -1,7 +1,0 @@
-import { Test } from 'src/entities/test';
-
-export class TestRepository {
-    public async getTest() {
-        return await Test.find();
-    }
-}
